# echango-api-main/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/cities:
    post:
      operationId: CitiesController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCityDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
      tags: &ref_3
        - Cities
    get:
      operationId: CitiesController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationCityResponseDto'
      tags: *ref_3
  /api/v1/cities/search:
    get:
      operationId: CitiesController_searchByName_v1
      parameters:
        - name: name
          required: true
          in: query
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationCityResponseDto'
      tags: *ref_3
  /api/v1/cities/all-or-search:
    get:
      operationId: CitiesController_getAllOrSearch_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: name
          required: true
          in: query
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationCityResponseDto'
      tags: *ref_3
  /api/v1/cities/{id}:
    get:
      operationId: CitiesController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
      tags: *ref_3
    patch:
      operationId: CitiesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCityDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
      tags: *ref_3
    delete:
      operationId: CitiesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      tags: *ref_3
...
  /api/v1/analytics/dashboard:
    get:
      operationId: AnalyticsController_getDashboard_v1
      parameters:
        - name: period
          required: true
          in: query
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security: &ref_17
        - bearer: []
      tags: &ref_18
        - Analytics
  /api/v1/analytics/track/profile-view/{officeId}:
    post:
      operationId: AnalyticsController_trackProfileView_v1
      parameters:
        - name: officeId
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '201':
          description: ''
      security: *ref_17
      tags: *ref_18
  /api/v1/analytics/track/phone-call/{officeId}:
    post:
      operationId: AnalyticsController_trackPhoneCall_v1
      parameters:
        - name: officeId
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackPhoneCallDto'
      responses:
        '201':
          description: ''
      security: *ref_17
      tags: *ref_18
  /api/v1/analytics/track/gps-click/{officeId}:
    post:
      operationId: AnalyticsController_trackGpsClick_v1
      parameters:
        - name: officeId
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '201':
          description: ''
      security: *ref_17
      tags: *ref_18
...
  /api/v1/auth/email/login:
    post:
      operationId: AuthController_login_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthEmailLoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
      tags: &ref_23
        - Auth
  /api/v1/auth/email/register:
    post:
      operationId: AuthController_register_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterLoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
      tags: *ref_23
  /api/v1/auth/email/confirm:
    post:
      operationId: AuthController_confirmEmail_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfirmEmailDto'
      responses:
        '204':
          description: ''
      tags: *ref_23
  /api/v1/auth/email/confirm/new:
    post:
      operationId: AuthController_confirmNewEmail_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfirmEmailDto'
      responses:
        '204':
          description: ''
      tags: *ref_23
  /api/v1/auth/forgot/password:
    post:
      operationId: AuthController_forgotPassword_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthForgotPasswordDto'
      responses:
        '204':
          description: ''
      tags: *ref_23
  /api/v1/auth/reset/password:
    post:
      operationId: AuthController_resetPassword_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResetPasswordDto'
      responses:
        '204':
          description: ''
      tags: *ref_23
  /api/v1/auth/change/password:
    post:
      operationId: AuthController_changePassword_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthChangePasswordDto'
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags: *ref_23
  /api/v1/auth/me:
    get:
      operationId: AuthController_me_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearer: []
      tags: *ref_23
    patch:
      operationId: AuthController_update_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearer: []
      tags: *ref_23
    delete:
      operationId: AuthController_delete_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags: *ref_23
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponseDto'
      security:
        - bearer: []
      tags: *ref_23
  /api/v1/auth/logout:
    post:
      operationId: AuthController_logout_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags: *ref_23
...
info:
  title: ExchangGo24 API
  description: ' API docs'
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    FileType:
      type: object
      properties:
        id:
          type: string
          example: cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae
        path:
          type: string
          example: https://example.com/path/to/file.jpg
      required:
        - id
        - path
    Role:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: admin
      required:
        - id
        - name
    Status:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: active
      required:
        - id
        - name
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          example: john.doe@example.com
        provider:
          type: string
          example: email
        socialId:
          type: string
          example: '1234567890'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        photo:
          $ref: '#/components/schemas/FileType'
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - provider
        - socialId
        - firstName
        - lastName
        - photo
        - role
        - status
        - createdAt
        - updatedAt
        - deletedAt
    WorkingHour:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the working hour record
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dayOfWeek:
          type: string
          description: Day of the week
          example: Monday
          enum:
            - Monday
            - Tuesday
            - Wednesday
            - Thursday
            - Friday
            - Saturday
            - Sunday
        isActive:
          type: boolean
          description: Indicates if the day is active for working hours
          example: true
        fromTime:
          type: string
          description: Start time of working hours
          example: '09:00'
          nullable: true
        toTime:
          type: string
          description: End time of working hours
          example: '17:00'
          nullable: true
        hasBreak:
          type: boolean
          description: Indicates if there is a break during this day
          example: false
        breakFromTime:
          type: string
          description: Start time of break
          example: '12:00'
          nullable: true
        breakToTime:
          type: string
          description: End time of break
          example: '13:00'
          nullable: true
        officeId:
          type: string
          description: ID of the associated office
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
          example: '2023-01-02T12:00:00Z'
      required:
        - id
        - dayOfWeek
        - isActive
        - fromTime
        - toTime
        - hasBreak
        - breakFromTime
        - breakToTime
        - officeId
        - createdAt
        - updatedAt
    Office:
      type: object
      properties:
        id:
          type: string
        owner:
          $ref: '#/components/schemas/User'
        officeName:
          type: string
        location:
          type: object
        registrationNumber:
          type: string
        currencyExchangeLicenseNumber:
          type: string
        address:
          type: string
        city:
          type: object
        state:
          type: object
        country:
          type: object
        primaryPhoneNumber:
          type: string
        secondaryPhoneNumber:
          type: object
        thirdPhoneNumber:
          type: object
        whatsappNumber:
          type: object
        logo:
          type: object
        mainImage:
          type: object
        slug:
          type: object
        isActive:
          type: boolean
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          type: object
        distanceInKm:
          type: number
        rates:
          type: array
          items:
            type: string
        equivalentValue:
          type: number
        images:
          type: object
        workingHours:
          type: array
          items:
            type: string
        email:
          type: object
        todayWorkingHours:
          $ref: '#/components/schemas/WorkingHour'
      required:
        - id
        - owner
        - officeName
        - location
        - registrationNumber
        - currencyExchangeLicenseNumber
        - address
        - city
        - state
        - country
        - primaryPhoneNumber
        - secondaryPhoneNumber
        - thirdPhoneNumber
        - whatsappNumber
        - logo
        - mainImage
        - slug
        - isActive
        - isVerified
        - isFeatured
        - createdAt
        - updatedAt
        - deletedAt
        - distanceInKm
        - rates
        - equivalentValue
        - images
        - workingHours
        - email
        - todayWorkingHours
...
    FileDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id
...
    Country:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unicode:
          type: string
        emoji:
          type: string
        alpha2:
          type: string
        dialCode:
          type: string
        region:
          type: string
        capital:
          type: string
        alpha3:
          type: string
      required:
        - id
        - name
        - unicode
        - emoji
        - alpha2
        - dialCode
        - region
        - capital
        - alpha3
    City:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          $ref: '#/components/schemas/Country'
        office:
          $ref: '#/components/schemas/Office'
        offices:
          description: Array of offices in this city
          type: array
          items:
            $ref: '#/components/schemas/Office'
        numberOfOffices:
          type: number
          description: Number of offices in this city
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - country
        - office
        - offices
        - numberOfOffices
        - createdAt
        - updatedAt
...
    CountryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unicode:
          type: string
        emoji:
          type: string
        alpha2:
          type: string
        dialCode:
          type: string
        region:
          type: string
        capital:
          type: string
        alpha3:
          type: string
      required:
        - id
        - name
        - unicode
        - emoji
        - alpha2
        - dialCode
        - region
        - capital
        - alpha3
    CreateCityDto:
      type: object
      properties:
        name:
          type: string
          example: Rabat
        country:
          example: MA
          allOf:
            - $ref: '#/components/schemas/CountryDto'
      required:
        - name
        - country
    InfinityPaginationCityResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/City'
        hasNextPage:
          type: boolean
          example: true
      required:
        - data
        - hasNextPage
    UpdateCityDto:
      type: object
      properties:
        name:
          type: string
          example: Rabat
        country:
          example: MA
          allOf:
            - $ref: '#/components/schemas/CountryDto'
...
    TrackPhoneCallDto:
      type: object
      properties:
        phoneNumber:
          type: string
          example: Phone Number
        phoneType:
          type: string
          example: Phone Type
      required:
        - phoneNumber
        - phoneType
...
    AuthEmailLoginDto:
      type: object
      properties:
        email:
          type: string
          example: test1@example.com
        password:
          type: string
      required:
        - email
        - password
    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        tokenExpires:
          type: number
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - refreshToken
        - tokenExpires
        - user
    AuthRegisterLoginDto:
      type: object
      properties:
        email:
          type: string
          example: test1@example.com
        password:
          type: string
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - email
        - password
        - firstName
        - lastName
    AuthConfirmEmailDto:
      type: object
      properties:
        hash:
          type: string
      required:
        - hash
    AuthForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: test1@example.com
      required:
        - email
    AuthResetPasswordDto:
      type: object
      properties:
        password:
          type: string
        hash:
          type: string
      required:
        - password
        - hash
    AuthChangePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: currentPassword123
          description: Current password of the user
        newPassword:
          type: string
          example: newPassword123
          description: New password for the user
        confirmPassword:
          type: string
          example: newPassword123
          description: Confirmation of the new password
      required:
        - oldPassword
        - newPassword
        - confirmPassword
    RefreshResponseDto:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        tokenExpires:
          type: number
      required:
        - token
        - refreshToken
        - tokenExpires
    AuthUpdateDto:
      type: object
      properties:
        photo:
          $ref: '#/components/schemas/FileDto'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: new.email@example.com
        password:
          type: string
        oldPassword:
          type: string
...